Conditional query matching
This addition to api-mock allows conditional matching of request bodies in examples. In its simplest form,
it will determine which request payload (body+header combo) matches the request received through express, and
return the response payload from it.
If it fails to find one, it returns the last defined payload.

In addition to this, it still fully honours Prefer header, providing backward-compatibility with the original
api-mock.

Use this if you would like to define routes such as

# /v5/test{?foo,bar}

## Test [GET]

+ Request

    + Body

         {
             foo: bar
         }

+ Response 200 (application/json)

    +
    + Body

        { baz: true}

+ Response 404 (application/json)